+++ docs/changelog.rst
@@ -20,6 +20,7 @@ Fixed
 .....
 
 * Add the code to deal with tag-by-uuid in the streams/viewsets.py and streams/urls/api.py. (`#607 <https://git.feneas.org/socialhome/socialhome/issues/607>`_)
+* Fallback to default locale (en-us) when a locale is not supported in Arrow.
 
 0.13.0 (2021-08-15)
 -------------------